Updates dependencies.
Changes make the new crypto format incompatible with the previous one.
This is probably something to keep an eye out for, as it can render
stores inaccessible. The new library used for crypto, however, is more
unified than 2 used before, said-to-be more secure and is some sort of
default for the rust ecosystem, so this problem may not repeat itself.